+++ src/components/Edit9LineDialog.js
@@ -1,0 +1,139 @@
+import React from 'react'
+
+import { LatLon } from 'geodesy/mgrs'
+
+import Button from '@material-ui/core/Button'
+import Dialog from '@material-ui/core/Dialog'
+import DialogActions from '@material-ui/core/DialogActions'
+import DialogContent from '@material-ui/core/DialogContent'
+import DialogContentText from '@material-ui/core/DialogContentText'
+import DialogTitle from '@material-ui/core/DialogTitle'
+import { makeStyles } from '@material-ui/core/styles'
+import TextField from '@material-ui/core/TextField'
+
+const useStyles = makeStyles(theme => ({
+  dialog: {
+    padding: theme.spacing(2),
+  },
+}))
+
+export default (props) => {
+  const classes = useStyles()
+
+  const [label, setLabel] = React.useState(props.marker.data === null ? '' : props.marker.data.label)
+  const [typeMethod, setTypeMethod] = React.useState(props.marker.data === null ? '' : props.marker.data.typeMethod)
+  const [ipHdgDistance, setIpHdgDistance] = React.useState(props.marker.data === null ? 'N/A' : props.marker.data.ipHdgDistance)
+  const [elevation, setElevation] = React.useState(props.marker.data === null ? props.marker.elevation : props.marker.data.elevation)
+  const [description, setDescription] = React.useState(props.marker.data === null ? props.marker.title : props.marker.data.description)
+  const [location, setLocation] = React.useState(props.marker.data === null ? LatLon.parse(props.marker.latlng.lat, props.marker.latlng.lng).toUtm().toMgrs().toString() : props.marker.data.location)
+  const [mark, setMark] = React.useState(props.marker.data === null ? 'None' : props.marker.data.mark)
+  const [friendlies, setFriendlies] = React.useState(props.marker.data === null ? '' : props.marker.data.friendlies)
+  const [egress, setEgress] = React.useState(props.marker.data === null ? '' : props.marker.data.egress)
+  const [remarks, setRemarks] = React.useState(props.marker.data === null ? '' : props.marker.data.remarks)
+
+  const handleSubmit = () => {
+    const data = {
+      type: '9line',
+      label: label,
+      typeMethod: typeMethod,
+      ipHdgDistance: ipHdgDistance,
+      elevation: elevation,
+      description: description,
+      location: location,
+      mark: mark,
+      friendlies: friendlies,
+      egress: egress,
+      remarks: remarks
+    }
+
+    props.submit(props.marker, data)
+  }
+
+  return (
+    <Dialog
+      className={classes.dialog}
+      open={props.open}
+      onClose={props.toggle}
+    >
+      <DialogTitle>Edit {props.marker.title} 9-Line</DialogTitle>
+      <DialogContent>
+        <TextField
+          autoFocus={true}
+          margin='dense'
+          label='Label'
+          onChange={event => setLabel(event.target.value)}
+          fullWidth
+          value={label}
+        />
+        <TextField
+          margin='dense'
+          label='Type / Method'
+          onChange={event => setTypeMethod(event.target.value)}
+          fullWidth
+          value={typeMethod}
+        />
+
+        <TextField
+          margin='dense'
+          label='IP / HDG / Distance'
+          onChange={event => setIpHdgDistance(event.target.value)}
+          fullWidth
+          value={ipHdgDistance}
+        />
+        <TextField
+          margin='dense'
+          label='Elevation'
+          onChange={event => setElevation(event.target.value)}
+          fullWidth
+          value={elevation}
+        />
+        <TextField
+          margin='dense'
+          label='Description'
+          onChange={event => setDescription(event.target.value)}
+          fullWidth
+          value={description}
+        />
+        <TextField
+          margin='dense'
+          label='Location'
+          onChange={event => setLocation(event.target.value)}
+          fullWidth
+          value={location}
+        />
+        <TextField
+          margin='dense'
+          label='Mark'
+          onChange={event => setMark(event.target.value)}
+          fullWidth
+          value={mark}
+        />
+        <TextField
+          margin='dense'
+          label='Friendlies'
+          onChange={event => setFriendlies(event.target.value)}
+          fullWidth
+          value={friendlies}
+        />
+        <TextField
+          margin='dense'  
+          label='Egress'
+          onChange={event => setEgress(event.target.value)}
+          fullWidth
+          value={egress}
+        />
+        <TextField
+          margin='dense'
+          label='Remarks &amp; Restrictions'
+          onChange={event => setRemarks(event.target.value)}
+          fullWidth
+          value={remarks}
+        />
+      </DialogContent>
+      <DialogActions>
+        <Button onClick={handleSubmit} color='primary'>Save Changes</Button>
+        <Button onClick={props.toggle}>Cancel</Button>
+      </DialogActions>
+    </Dialog>
+  )
+}